move some controller
