+++ packages/Webkul/Admin/src/Http/Controllers/Settings/LocaleController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\Admin\Http\Controllers\Core;
+namespace Webkul\Admin\Http\Controllers\Settings;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 use Webkul\Admin\Http\Controllers\Controller;
