+++ packages/Webkul/Admin/src/Http/Controllers/Settings/TaxRateController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Webkul\Admin\Http\Controllers\Tax;
+namespace Webkul\Admin\Http\Controllers\Settings;
 
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Validator;
