+++ packages/Webkul/Admin/src/Routes/notification-routes.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use Webkul\Admin\Http\Controllers\Notification\NotificationController;
+use Webkul\Admin\Http\Controllers\NotificationController;
 
 /**
  * Notification routes.
